+++ tests/unit_tests/_primaite/_simulator/_system/_services/test_ftp.py
@@ -3,6 +3,8 @@ from ipaddress import IPv4Address
 import pytest
 
 from primaite.simulator.network.hardware.base import Node
+from primaite.simulator.network.hardware.nodes.computer import Computer
+from primaite.simulator.network.hardware.nodes.server import Server
 from primaite.simulator.network.protocols.ftp import FTPCommand, FTPPacket
 from primaite.simulator.network.transmission.network_layer import IPProtocol
 from primaite.simulator.network.transmission.transport_layer import Port
@@ -12,7 +14,9 @@ from primaite.simulator.system.services.ftp.ftp_server import FTPServer
 
 @pytest.fixture(scope="function")
 def ftp_server() -> Node:
-    node = Node(hostname="ftp_server")
+    node = Server(
+        hostname="ftp_server", ip_address="192.168.1.10", subnet_mask="255.255.255.0", default_gateway="192.168.1.1"
+    )
     node.software_manager.install(software_class=FTPServer)
     node.software_manager.software["FTPServer"].start()
     return node
@@ -20,7 +24,9 @@ def ftp_server() -> Node:
 
 @pytest.fixture(scope="function")
 def ftp_client() -> Node:
-    node = Node(hostname="ftp_client")
+    node = Computer(
+        hostname="ftp_client", ip_address="192.168.1.11", subnet_mask="255.255.255.0", default_gateway="192.168.1.1"
+    )
     return node